+++ blayers/__init__.py
@@ -2,6 +2,7 @@ from importlib.metadata import PackageNotFoundError, version
 
 from blayers.decorators import autoreparam, autoreshape
 from blayers.fit import FittedModel, Predictions, fit, sample_prior
+from blayers.latex import model_to_latex
 from blayers.layers import (
     AdaptiveLayer,
     BilinearLayer,
@@ -86,6 +87,7 @@ __all__ = [
     # Fit
     "fit",
     "sample_prior",
+    "model_to_latex",
     "FittedModel",
     "Predictions",
 ]